[3/n] Pin ActivityStack
Implementing pinTopActivityStack API, which required in extensions
vendor API level 4.

This is the initial change to make the top-most TF container to
split with another TF container below.

Also adding a SplitPinContainer to hold the two TF containers.
The primary container is set to mutable which could be changed
and updated in runtime (in latter CLs).

The pinned TF container and SplitPinContainer is ensured to be on
top on the client side. The server side change will be done in the
following CLs.

Bug: 208573140
Test: atest SplitControllerTest
Change-Id: Ic932756f0adb4997b3729f0ed53dd9542dd4b7ae