Close system dialogs as Escape fallback
At the moment the Escape key won't do anything most of the time. We considered adding a
fallback to the back key, but that could cause unintended side effects.

We have a intent, close system dialogs, that was clamped down as "too powerful" for apps,
and currently used by the test instrumentation. This seems like the perfect option for Escape.

Bug: 278851142
Test: flashed on device
Change-Id: I6ebf22be96313f09a94012f12e75a94df57a29c4