Prevent UserSwitchingDialog flashing when setup wizard is shown.
When switching to a newly created user, or switching to a user with
an uncompleted setup, setup wizard (SUW) is shown during the user
switch and SUW is drawn over the user switching dialog. Then when
the user switching dialog is dismissed it becomes visible again for a
moment.

This CL prevents that jank by freezing the screen while showing the
user switching dialog, if the target user has not completed the SUW.

Bug: 279773661
Test: Visual
Change-Id: Ia611d3d062ba78c8def173ab177f183d5494c919