Do not apply alpha if window is trusted.
Non-trusted overlays with FLAG_NOT_TOUCHABLE
have an alpha auto-applied to them so that
it's guaranteed you can see the window below
it. Before this change we checked if the window
session had the trusted overlay flag set, if
it was we would not apply the alpha.

However, we recently updated window sessions
to be considered a trusted overlay if they
have the trusted overlay flag or if they
are a system application overlay. This change
was not reflected for the automatic alpha
applied. This change updates the check to
the isTrustedOverlay method on the window
session so that system application overlays
are allowed to set their own alpha to
FLAG_NOT_TOUCHABLE overlays.

Test: Manually confirmed SysUI which has
the system application overlay permission
is able to set its own alpha on a
FLAG_NOT_TOUCHABLE overlay.

Bug: 274789900
Change-Id: Ia01052cc1e52ceece8a27f13ca3f8e9136a32b99