Allow setting REVIEW_REQUIRED in updatePermissionFlags().
So that CTS tests can easily both set and clear the permission flag
when testing. Note that the existing check for SHELL_UID doesn't work
for our tests, because the calling UID will remain the test app's own
UID even if it uses adoptShellPermissionIdentity().

The current comment in PermissionManagerServiceImpl did intend to
limit the ability to change the REVIEW_REQUIRED flag to
shell/system/root, however the code actually allows remove the
flag (while disallowing setting it), because it only drops the flag
from flag values but not flag mask.

This doesn't make sense because setting the flag is similar to
revoking the permission, but removing the flag is similar to granting
the permission, so it's allowing the more dangerous action but
disallowing the more restrictive action.

The updatePermissionFlags() method already requires the heavily locked
down GRANT/REVOKE_RUNTIME_PERMISSIONS to call, so removing the
requirement in its current state won't likely allow any new malicious
party to temper with the flag - as for entities that have
GRANT/REVOKE_RUNTIME_PERMISSIONS, they can do much worse things and
being allowed to add this flag won't add to their power.

Meanwhile, this REVIEW_REQUIRED flag is only for pre-M apps, but after
all these years, and with store policy and the low target SDK block,
there are much less pre-M apps to support anyway.

Bug: 263504888
Test: presubmit
Change-Id: I61f328a65cd3871dbac27283bd666e7b858842ba