Update DisplayMetrics.densityDpi Javadoc to refer to all DENSITY_ constants
The immediate motive for getting around to updating this is the text
"may be backported".  According to Dianne's recollection, "we had in the
past allowed OEMs to use new densities we have defined on platform
versions prior to when they were introduced": this documents that
possibility and clarifies that this might happen again.

See related change aosp/2579165 "Backport DENSITY_520 to
android13-tests-dev"

Test: N/A
Bug: 274567391

Change-Id: I3bc25630eab0208c83837b74c53facea637169d6