[Sb refactor] Ensure config is fetched on flow start
In the case where the `CARRIER_CONFIG_CHANGED` broadcast is sent out
before the default data sub RAT config is read, it will read the
system-defined default carrier config instead of the appropriate one for
the current default data sub.

This change is simple: make sure to always fetch the system current
carrier config when the flow starts.

Bug: 280390475
Test: MobileConnectionsRepositoryTest
Change-Id: I868a840ee8465b0313bc85e5c35c0d40ca9d509a