fix(non linear font scaling)!: fix QS text being cut off when 200% font scaling
Originally the QS tile height is fixed to 80dp and may contain 2 textViews in tile content. When font scaling changed to 200%, the height from 80dp is not enough for the textViews. Therefore, in TileLayout, we estimate the needed height for the tile under given font scaling, and compare with the height from resources/dimens. The larger height would become the exact tile height, to have a proper tile height for full tile content view.

Bug: b/272072639
Test: manually - attached video in bug
      atest TileLayoutTest
      atest QuickQSPanelTest
Change-Id: I6054ef1a6fd0fc28522181f5bc7774cb0abaf14a