SlicePurchaseActivity handle user data based on contents type
If contents type is not present, we need to append user data to the url
and send it as a GET request. If contents type is xml or json, we need
to send the user data as a POST request. If user data is encoded, we
need to decode it before sending it in the POST request. If the contents
type is specified but user data does not exist, return an error.

Test: atest CarrierDefaultAppUnitTests
Test: manual test userdata is properly appended or posted
Bug: 282905562
Change-Id: I59e1e28e7e1bd583307da55187886b8bb0798006
Merged-In: I59e1e28e7e1bd583307da55187886b8bb0798006