Changed VibrationAttributes of Unfold Haptics
There was feedback that the haptic when unfolding was too subtle, so we are increasing the strength to make it more prominent.

In this change, USAGE_HARDWARE_FEEDBACK was added as a VibrationAttribute for the unfold haptics.
Previously, when no VibrationAttributes were specified, it defaulted to USAGE_UNKNOWN, which meant that the haptics were controlled by the 'Media Vibration' user setting slider.
By associating it with USAGE_HARDWARE_FEEDBACK, the haptics will now be controlled by the 'Touch Feedback' user setting slider.

Test: UnfoldHapticsPlayerTest
Bug: 281642496
Relevant Logs: https://paste.googleplex.com/6294954983817216
Change-Id: I3241a924bf843e590e7d947b5375cd69cda73139